requirements.txt
flask==3.1.2
opencv-python==4.12.0.88
mysql-connector-python==9.5.0
numpy==2.2.6
pillow==12.0.0